Vincular carro com usuario

+++ src/database/VehiclesDatabase.jsx
@@ -8,6 +8,7 @@ export const initVehiclesDB = () => {
         tx.executeSql(
             `CREATE TABLE IF NOT EXISTS vehicles (
                 id INTEGER PRIMARY KEY AUTOINCREMENT,
+                userId INTEGER,
                 name TEXT,
                 brand TEXT,
                 model TEXT,
@@ -28,12 +29,12 @@ export const initVehiclesDB = () => {
 };
 
 {/* Inserir Novo Veículo */}
-export const insertVehicle = (vehicle) => {
+export const insertVehicle = (vehicle, userId) => {
     db.transaction(tx => {
         tx.executeSql(
-            `INSERT INTO vehicles (name, brand, model, version, color, manufactureYear, licensePlate, fuelType, transmission, engine, mileage)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);`,
-            [vehicle.name, vehicle.brand, vehicle.model, vehicle.version, vehicle.color, vehicle.manufactureYear, vehicle.licensePlate, vehicle.fuelType, vehicle.transmission, vehicle.engine, vehicle.mileage],
+            `INSERT INTO vehicles (userId, name, brand, model, version, color, manufactureYear, licensePlate, fuelType, transmission, engine, mileage)
+            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);`,
+            [userId, vehicle.name, vehicle.brand, vehicle.model, vehicle.version, vehicle.color, vehicle.manufactureYear, vehicle.licensePlate, vehicle.fuelType, vehicle.transmission, vehicle.engine, vehicle.mileage],
             (_, resultSet) => console.log('Veículo inserido com sucesso:', resultSet),
             (_, error) => console.log('Erro ao inserir veículo:', error)
         );
@@ -54,12 +55,26 @@ export const fetchVehicles = () => {
     });
 };
 
+{/* Carregar Veículos por Usuario */}
+export const fetchUserVehicles = (userId) => {
+    return new Promise((resolve, reject) => {
+        db.transaction(tx => {
+            tx.executeSql(
+                `SELECT * FROM vehicles WHERE userId = ?;`,
+                [userId],
+                (_, { rows: { _array } }) => resolve(_array),
+                (_, error) => reject(error)
+            );
+        });
+    });
+};
+
 {/* Atualiizar Veículo */}
-export const updateVehicle = (vehicle) => {
+export const updateVehicle = (vehicle, userId) => {
     db.transaction(tx => {
         tx.executeSql(
-            `UPDATE vehicles SET name = ?, brand = ?, model = ?, version = ?, color = ?, manufactureYear = ?, licensePlate = ?, fuelType = ?, transmission = ?, engine = ?, mileage = ? WHERE id = ?;`,
-            [vehicle.name, vehicle.brand, vehicle.model, vehicle.version, vehicle.color, vehicle.manufactureYear, vehicle.licensePlate, vehicle.fuelType, vehicle.transmission, vehicle.engine, vehicle.mileage, vehicle.id],
+            `UPDATE vehicles SET name = ?, brand = ?, model = ?, version = ?, color = ?, manufactureYear = ?, licensePlate = ?, fuelType = ?, transmission = ?, engine = ?, mileage = ? WHERE id = ? AND userId = ?;`,
+            [vehicle.name, vehicle.brand, vehicle.model, vehicle.version, vehicle.color, vehicle.manufactureYear, vehicle.licensePlate, vehicle.fuelType, vehicle.transmission, vehicle.engine, vehicle.mileage, vehicle.id, userId],
             (_, resultSet) => console.log('Veículo atualizado com sucesso:', resultSet),
             (_, error) => console.log('Erro ao atualizar veículo:', error)
         );
@@ -67,13 +82,33 @@ export const updateVehicle = (vehicle) => {
 };
 
 {/* Deletar Veículo */}
-export const deleteVehicle = (vehicleId) => {
+export const deleteVehicle = (vehicleId, userId) => {
     db.transaction(tx => {
         tx.executeSql(
-            `DELETE FROM vehicles WHERE id = ?;`,
-            [vehicleId],
+            `DELETE FROM vehicles WHERE id = ? AND userId = ?;`,
+            [vehicleId, userId],
             (_, resultSet) => console.log('Veículo excluído com sucesso:', resultSet),
             (_, error) => console.log('Erro ao excluir veículo:', error)
         );
     });
 };
+
+{/* Resetar Banco de Dados de Usuarios */}
+export const deleteAllVehicles = () => {
+    return new Promise((resolve, reject) => {
+       db.transaction(tx => {
+         tx.executeSql(
+           `DROP TABLE vehicles;`,
+           [],
+           () => {
+             console.log('Todos os veículos foram apagados com sucesso');
+             resolve();
+           },
+           (_, error) => {
+             console.log('Erro ao apagar veículos:', error);
+             reject(error);
+           }
+         );
+       });
+    });
+   };
